Moved db to utils
