+++ api/src/db_loaded_items.js
@@ -15,7 +15,7 @@
 */
 
 import { logger } from "./utils/logging.js";
-import { db_conn, db_check } from "./models/index.js";
+import { db_conn, db_check } from "./utils/db.js";
 
 let committee_display_to_id;
 let committee_id_to_display;
