+++ api/src/models/access.js
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-import { db_conn } from "./index.js";
+import { db_conn } from "../utils/db.js";
 import { ACCESS_LEVEL } from "../common_items.js";
 
 async function checkApprovalExists(id, committee, checkTreasurer=false) {
