+++ api/src/models/budgets.js
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-import { db_conn } from "./index.js";
+import { db_conn } from "../utils/db.js";
 
 async function clearBudget(comm, year) {
     return db_conn.promise().execute(
