+++ api/src/models/purchase.js
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-import { db_conn } from "./index.js";
+import { db_conn } from "../utils/db.js";
 import { ACCESS_LEVEL, max_fiscal_year_count } from "../common_items.js";
 
 async function getFullPurchaseByID(id) {
